refactor(llm): extract logprob confidence helpers to _logprob_confidence
Moves the logprob-driven description confidence layer (9 functions
plus the _BOILERPLATE_TOKENS sentinel) out of provider.py:
- _lp_token_text, _lp_token_logprob, _is_value_token (token utilities)
- _description_value_spans (regex span detection)
- _weighted_score_for_spans (per-span weighted score)
- logprob_confidence_score, _logprob_token_spans,
  logprob_confidence_score_for_text, confidence_from_logprobs
  (public scorers used by orchestrator + agents)

All callers continue to import from amx.llm.provider via re-export.
provider.py: 1401 -> 1244 LOC (-157, -11%).
